客户主机编号，出库必填。还有一个自定义字段，可填可不填2

+++ src/main/java/com/truck/service/impl/CartServiceImpl.java
@@ -63,9 +63,9 @@ public class CartServiceImpl implements ICartService {
     }
 
     public ServerResponse<CartVo> update(Integer adminId, Integer count, Integer stockId, BigDecimal cartPrice,String defineSn,String defineStr,String defineModelNo) {
-        if (count == null || stockId == null)
+        if (stockId == null)
             return ServerResponse.createByErrorCodeMessage(ResponseCode.ILLEGAL_ARGUMENT.getCode(), ResponseCode.ILLEGAL_ARGUMENT.getDesc());
-        if (count < 1)
+        if (count !=null && count < 1)
             return ServerResponse.createByErrorMessage("");
         Cart cart = cartMapper.selectCartByAdminIdStockId(adminId, stockId);
         if (cart != null){
@@ -84,11 +84,16 @@ public class CartServiceImpl implements ICartService {
             if (StringUtils.isNotBlank(defineModelNo)) {
                 cart.setDefineModelNo(defineModelNo);
             }
-            Integer real= cart.getAmount()+count;
-            if(real <=0){
-                cartMapper.deleteByPrimaryKey(cart.getCartId());
+
+            if(count !=null){
+                Integer real= cart.getAmount()+count;
+                if(real <=0){
+                    cartMapper.deleteByPrimaryKey(cart.getCartId());
+                }else{
+                    cart.setAmount(count);
+                    cartMapper.updateByPrimaryKeySelective(cart);
+                }
             }else{
-                cart.setAmount(count);
                 cartMapper.updateByPrimaryKeySelective(cart);
             }
         }
